src/test/java/edu/hw1/Task8Test.java
package edu.hw1;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;
import java.util.stream.Stream;
import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;

public class Task8Test {
    private static Stream<Arguments> provideParameters() {
        return Stream.of(
            Arguments.of(new int[][] {
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1}
            }, false),
            Arguments.of(new int[][] {
                {0, 0, 0, 1, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 1, 0, 0, 0, 1, 0, 0},
                {0, 0, 0, 0, 1, 0, 1, 0},
                {0, 1, 0, 0, 0, 1, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 1, 0, 0, 0, 0, 0, 1},
                {0, 0, 0, 0, 1, 0, 0, 0}
            }, true),
            Arguments.of(new int[][] {
                {1, 0, 1, 0, 1, 0, 1, 0},
                {0, 1, 0, 1, 0, 1, 0, 1},
                {0, 0, 0, 0, 1, 0, 1, 0},
                {0, 0, 1, 0, 0, 1, 0, 1},
                {1, 0, 0, 0, 1, 0, 1, 0},
                {0, 0, 0, 0, 0, 1, 0, 1},
                {1, 0, 0, 0, 1, 0, 1, 0},
                {0, 0, 0, 1, 0, 1, 0, 1}
            }, false),
            Arguments.of(new int[][] {
                {0, 0, 0, 0, 1, 0, 0, 0},
                {0, 0, 0, 0, 0, 1, 0, 0},
                {0, 0, 0, 1, 0, 0, 0, 0},
                {1, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 1, 0, 0, 0},
                {0, 0, 0, 0, 0, 1, 0, 0},
                {0, 0, 0, 0, 0, 1, 0, 0},
                {1, 0, 0, 0, 0, 0, 0, 0}
            }, false)
        );
    }

    @ParameterizedTest
    @MethodSource("provideParameters")
    public void isNestable_InputTwoArrays_ResultCanFirstArrayNestedInSecondArray(
        int[][] inputBoard,
        boolean expectedKnightBoardCapture
    ) {
        // When
        boolean actualKnightBoardCapture = Task8.knightBoardCapture(inputBoard);

        // Then
        assertThat(actualKnightBoardCapture).isEqualTo(expectedKnightBoardCapture);
    }

    private static Stream<Arguments> provideParametersForIllegalArgumentException() {
        return Stream.of(
            Arguments.of(new int[][] {
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1},
                {1, 1, 1, 1},
                {1, 1, 1},
                {1, 1},
                {1}
            }, "The board should be 8x8"),
            Arguments.of(new int[][] {
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 999999999, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1}
            }, "All elements on board should be equals 0 or 1"),
            Arguments.of(new int[][] {
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, -999999999, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1, 1}
            }, "All elements on board should be equals 0 or 1"),
            Arguments.of(new int[][] {
                {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1, 1},
                {1, 1, 1, 1, 1},
                {1, 1, 1, 1},
                {1, 1, 1},
                {1, 1},
                {1}
            }, "The board should be 8x8"),
            Arguments.of(new int[][] {
                {1, 1, 1, 1},
                {1, 1, 1},
                {1, 1},
                {1}
            }, "The board should be 8x8"),
            Arguments.of(new int[][] {
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0}
            }, "No knights on the board")
        );
    }

    @ParameterizedTest
    @MethodSource("provideParametersForIllegalArgumentException")
    public void isNestable_InputTwoArrays_ResultIllegalArgumentException(
        int[][] inputBoard,
        String expectedMessage
    ) {
        assertThatThrownBy(() -> {
            Task8.knightBoardCapture(inputBoard);
        }).isInstanceOf(IllegalArgumentException.class)
            .hasMessage(expectedMessage);
    }

    private static Stream<Arguments> provideParametersForNullPointerException() {
        return Stream.of(
            Arguments.of(new int[][] {
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                null,
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0}
            }, NullPointerException.class),

            Arguments.of(null, NullPointerException.class)
        );
    }

    @ParameterizedTest
    @MethodSource("provideParametersForNullPointerException")
    public void isNestable_InputTwoArrays_ResultNullPointerException(
        int[][] inputBoard
    ) {
        assertThatThrownBy(() -> {
            Task8.knightBoardCapture(inputBoard);
        }).isInstanceOf(NullPointerException.class);
    }
}
